removed 4 warnings - redundanto .toInt method call

+++ src/main/scala/beam/utils/scenario/ScenarioLoader.scala
@@ -152,7 +152,7 @@ class ScenarioLoader(
             .find(_.vehicleCategory == VehicleCategory.Bike)
             .getOrElse(BeamVehicleType.defaultBikeBeamVehicleType)
         )
-        initialVehicleCounter += householdInfo.cars.toInt
+        initialVehicleCounter += householdInfo.cars
         totalCarCount += vehicleTypes.count(_.vehicleCategory.toString == "Car")
         val vehicleIds = new java.util.ArrayList[Id[Vehicle]]
         vehicleTypes.foreach { beamVehicleType =>
@@ -182,8 +182,8 @@ class ScenarioLoader(
     beamServices.beamConfig.beam.agentsim.agents.vehicles.downsamplingMethod match {
       case "SECONDARY_VEHICLES_FIRST" =>
         val rand = new Random(beamServices.beamConfig.matsim.modules.global.randomSeed)
-        val hh_car_count = collection.mutable.Map(households.groupBy(_.cars.toInt).toSeq: _*)
-        val totalCars = households.foldLeft(0)(_ + _.cars.toInt)
+        val hh_car_count = collection.mutable.Map(households.groupBy(_.cars).toSeq: _*)
+        val totalCars = households.foldLeft(0)(_ + _.cars)
         val goalCarTotal = math
           .round(beamServices.beamConfig.beam.agentsim.agents.vehicles.fractionOfInitialVehicleFleet * totalCars)
           .toInt
@@ -217,7 +217,7 @@ class ScenarioLoader(
         households.foreach { household =>
           nVehiclesOut += drawFromBinomial(
             rand,
-            household.cars.toInt,
+            household.cars,
             beamServices.beamConfig.beam.agentsim.agents.vehicles.fractionOfInitialVehicleFleet
           )
         }
